+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [UNRELEASED] · YYYY-MM-DD
 ### ✨ Added
 - Add tooltips in many places ([#694](https://github.com/cbrnr/mnelab/pull/694) by [Clemens Brunner](https://github.com/cbrnr))
+- Add History entries for editing events or annotations ([#117](https://github.com/cbrnr/mnelab/issues/117) by [Clemens Brunner](https://github.com/cbrnr))
 
 ### 🔧 Fixed
 - Apply the selected Number of Components setting when running ICA ([#694](https://github.com/cbrnr/mnelab/pull/694) by [Clemens Brunner](https://github.com/cbrnr))